requirements.txt

requests
xmltodict
python-dotenv
